# Vectorize StateMachine::ProcessString